+++ R/color_annotate.R
@@ -5,11 +5,10 @@
 #' @author George Moroz <agricolamz@gmail.com>
 #' @param gloss vector of characters.
 #' @export
-#' @importFrom rmarkdown metadata
+#' @importFrom knitr is_latex_output
 
 color_annotate <- function(gloss){
-  if(!is.null(rmarkdown::metadata$output) &&
-     grepl("latex", unlist(rmarkdown::metadata$output))){
+  if(knitr::is_latex_output()){
     paste(paste0('\\colorbox{cyan}{', gloss, '}'))
   } else {
     paste(paste0('<span style="background-color:#00d5ff;">', gloss, '</span>'))